+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.0 - 2015-07-21
 
 - Implement `update()` (fixes [#5](https://github.com/urish/firebase-server/issues/5))
+- Implement `transaction()`
 - Bugfix: `remove()` triggers two value events (fixes [#6](https://github.com/urish/firebase-server/issues/6))
 
 ## 0.2.0 - 2015-06-12